添加了主机的EEPROM

+++ 程序/slave/SYSTEM/sys.h
@@ -1,16 +1,16 @@
 #ifndef _SYS_H_
 #define _SYS_H_
 
-#include  <INTRINS.H>
+#include "intrins.h"
 #include "reg52.h"
-#include "delay.h"
-#include "usart.h"
+
 
 typedef unsigned char uchar;
 typedef unsigned int uint;
 
-typedef	unsigned char BYTE;
-typedef	unsigned short WORD;
+typedef unsigned char BYTE;
+typedef unsigned int WORD;
+
 
 #endif
